The biggest change for Db_Home_Page, was that the following code block was removed:  public static void main(String[] args) {         EventQueue.invokeLater(() -> {             try {                 Db_Sign_In frame = new Db_Sign_In();                 frame.setVisible(true);             } catch (Exception e) {                 e.printStackTrace();             }         });     } I tried to see what the use of this was but there was no change in the program functionality after removing the code block.
The buttons were updated to different locations in the frame, and I changed the buttons to only be "Yes" or "No" buttons instead of "Yes", "No", "Cancel buttons".

I am adding a new button to display an employee directory. The table and data still need to be added and created in the database.